requirements.txt
Django==3.2.8
django-import-export==2.6.1
django-pandas==0.6.4
pandas==1.3.3
plotly==5.3.1
numpy==1.21.2
